+++ src/rdbus/config/Output.cpp
@@ -1,4 +1,5 @@
 #include "Output.hpp"
+#include "Address.hpp"
 #include "utility.hpp"
 
 using namespace nlohmann;
@@ -8,20 +9,11 @@ namespace rdbus::config
 
 void from_json( const json& j, Output& x )
 {
-    constexpr int invalidPort = -1;
-    int port = invalidPort;
-    tools::parseKeyValue( j, "port", port );
-
-    std::string ip;
-    tools::parseKeyValue( j, "ip", ip );
-
     std::string type;
     tools::parseKeyValue( j, "type", type, "Missing 'type' field in 'output' section!" );
 
     tools::throwIf( type != "stdout" && type != "TCP/IP", "The only allowed output types are 'stdout' and 'TCP/IP'!" );
-    tools::throwIf( type == "stdout" && ( !ip.empty() || port != invalidPort ), "Multiple output types detected! Do not mix 'ip/port' with stdout 'type'!" );
-    tools::throwIf( type == "TCP/IP" && port == invalidPort, "No 'port' field!" );
-    tools::throwIf( type == "TCP/IP" && ip.empty(), "No 'ip' field!" );
+    tools::throwIf( type == "stdout" && j.contains( "address" ), "Multiple output types detected! Do not mix 'address' with stdout 'type'!" );
 
     if ( type == "stdout" )
     {
@@ -29,8 +21,9 @@ void from_json( const json& j, Output& x )
     }
     else if ( type == "TCP/IP" )
     {
-        x.ip = ip;
-        x.port = port;
+        Address address;
+        tools::parseKeyValue( j, "address", address, "Missing 'address' section for 'TCP/IP' output type!" );
+        x.address = address;
         x.type = Output::TCP_IP;
     }
 }