+++ src/rdbus/config/Output.hpp
@@ -1,18 +1,15 @@
 #pragma once
 
+#include "Address.hpp"
 #include <nlohmann/json.hpp>
 #include <optional>
-#include <string>
 
 namespace rdbus::config
 {
 
 struct Output
 {
-    using IP = std::string;
-    using Port = int;
-    std::optional< Port > port;
-    std::optional< IP > ip;
+    std::optional< Address > address;
 
     enum Type
     {